Move CPP example app into new structure
